requirements.txt
kafka-python
flask
apache-airflow-providers-mongo
pymongo
